D6P2 Solved

+++ 2023/day06/part2/src/competition/mod.rs
@@ -1,5 +1,5 @@
 #[derive(Debug, PartialEq, Eq)]
 pub struct Race {
-    pub start: u32,
-    pub duration: u32,
+    pub start: u64,
+    pub duration: u64,
 }
